fix(bridge): send Record timestamps at nanosecond resolution
The Bridge read only `msg.header.stamp.sec` and dropped `.nanosec`, then packed
it as the ingest protocol's plain integer time. Every Record's timestamp was
rounded to the second, so a Measurement polling faster than 1 Hz produced
Records that were indistinguishable in time — measured at 5 Hz: 8 Records,
3 distinct timestamps.

The wire format already supported the fix. Fluent Forward has an EventTime
extension (ext type 0x00, 4 bytes seconds + 4 nanoseconds), and the vendored
Vector 0.57.0 parses it with all nine digits intact — verified by feeding its
fluent source both frame shapes. Only the Bridge needed to change: Record now
carries seconds and nanoseconds, and the Forwarder packs EventTime.

Adds `epoch_nanos` as a new time_format, and makes it the default. `double`
cannot represent nanoseconds — a float64 has ~15-16 significant digits and
current epoch seconds spend 10 of them, so it tops out near microseconds. That
is IEEE 754, not an implementation limit. `double` and `iso8601` stay
available. Exactness also matters beyond precision: it is what makes the
timestamp usable as part of a Record identity, which #309 needs.

The schema follows: `date` becomes bigint in both init.sql files,
`to_timestamp(date)` becomes `to_timestamp(date / 1e9)` in the dashboard, and
params writing to those tables drop their explicit `time_format: "double"`.
Console-only demos keep theirs.

Regression cover: e2e_params.yaml's memory Measurement now polls at 5 Hz so
several Records land inside one second, and verify_zero_loss.py asserts a
nanosecond remainder exists, that some second holds 2+ Records (so it cannot
pass vacuously), and that no two Records share an instant. At 1 Hz a return to
whole-second stamps would not collide and would pass unnoticed.

Closes #308

+++ tools/e2e/sql/init.sql
@@ -6,8 +6,13 @@
 -- lands here. `tag` (added to every event by Vector's fluent source) + `date` (the
 -- normalized event timestamp, per the destination's time_key/time_format) are the
 -- verification key tools/e2e/scripts/verify_zero_loss.py uses.
+--
+-- `date` is bigint, not double precision: the default time_format is epoch_nanos, an
+-- exact integer count of nanoseconds since the epoch (#308). A double would round it
+-- back off below roughly microsecond resolution, which is what made timestamps useless
+-- for telling two Records of a fast Measurement apart.
 CREATE TABLE IF NOT EXISTS dc_records (
-  date double precision,
+  date bigint,
   tag text,
   group_key text,
   -- memory
@@ -39,7 +44,7 @@ CREATE TABLE IF NOT EXISTS dc_records (
 -- column set proven against real dc_bridge behavior in
 -- dc_bridge/dc_bridge_core/tests/uploader_tests.rs and dc_bridge/tests/end_to_end.rs.
 CREATE TABLE IF NOT EXISTS dc_files (
-  date double precision,
+  date bigint,
   kind text,
   group_name text,
   robot_name text,
